feat(caritas-theme): move new registration to core, cleanup styles & extensions
