+++ src/components/generateQrCode/generateQrCode.styles.scss
@@ -40,7 +40,7 @@
 
 			.text {
 				max-width: 500px;
-				margin-left: auto;
+				margin-left: 0;
 				margin-right: auto;
 			}
