+++ src/components/registration/metaInfo/MetaInfo.tsx
