+++ src/components/registration/registrationDataValidation.ts
