+++ src/components/registration/stepBar/StepBar.stories.ts
