+++ src/components/registration/stepBar/StepBar.tsx
