+++ src/components/registration/welcomeScreen/welcomeScreen.stories.ts
