+++ src/components/registration/welcomeScreen/welcomeScreen.styles.scss
@@ -1,0 +1,79 @@
+$welcome-screen-button-text-align: center !default;
+
+.registrationWelcome {
+	margin: 0;
+
+	h2 {
+		margin-bottom: $grid-base-two;
+	}
+
+	h4 {
+		line-height: 28px;
+		font-weight: 400;
+	}
+
+	& > * {
+		max-width: 500px;
+		margin: auto;
+	}
+
+	&__explanation,
+	&__buttonsWrapper {
+		margin: $grid-base-five auto 0;
+		max-width: 500px;
+		display: grid;
+		grid-template-columns: auto;
+		column-gap: $grid-base-three;
+		row-gap: $grid-base-three;
+
+		@include breakpoint($fromMedium) {
+			margin: $grid-base-six auto 0;
+			grid-template-columns: auto auto;
+			row-gap: $grid-base-four;
+			width: 100%;
+		}
+	}
+
+	&__buttonsWrapper {
+		border: 1px solid $border-default;
+		border-radius: $input-field-border-radius;
+		padding: $grid-base-four;
+
+		.text {
+			text-align: $welcome-screen-button-text-align;
+			margin-bottom: 14px;
+			color: $text-high-emphasis;
+			font-weight: 600;
+		}
+
+		.buttonWrapper {
+			text-align: center;
+		}
+
+		button {
+			min-width: auto;
+			width: 100%;
+		}
+
+		.button__wrapper {
+			&:focus {
+				outline: $focus-outline;
+				outline-offset: 4px;
+			}
+		}
+
+		&:focus:not(:focus-visible) {
+			outline: none;
+		}
+	}
+
+	&__consultingType {
+		font-size: $font-size-secondary;
+		font-weight: $font-weight-medium;
+		line-height: 14px;
+		letter-spacing: 1.5px;
+		text-transform: uppercase;
+		padding-bottom: $grid-base;
+		color: $text-low-emphasis;
+	}
+}
