+++ src/components/releaseNote/releaseNote.styles.scss
@@ -1,4 +1,8 @@
 .releaseNote {
+	.overay {
+		z-index: 9999;
+	}
+
 	.overlay__content {
 		text-align: left;
