+++ src/components/stage/stage.styles.scss
@@ -1,9 +1,43 @@
-$loginOverlayDelayDesktop: 1s;
-$loginOverlayDelayMobile: 2.5s;
+$animation-duration: 2.5s;
+$animation-duration-loader: $animation-duration * 0.25 * 3;
+$animation-duration-width: $animation-duration * 0.25;
+$animation-duration-title: calc($animation-duration / 7);
+
+$animation-easing-width: cubic-bezier(0, 0, 0.1, 1);
+$animation-easing-title: cubic-bezier(0, 0, 0.5, 1);
+
 $iconSize: 55px;
 $gridSpacing: $grid-base-four;
 
+.stageLayout {
+	@media (min-width: $fromXLarge) {
+		&__legalLinks {
+			display: none;
+		}
+	}
+
+	&__legalLinks {
+		justify-content: center;
+	}
+
+	&__legalLinksItem {
+		color: $tertiary;
+		text-transform: none;
+		font-weight: normal;
+
+		&:hover {
+			color: var(--skin-color-primary-hover, $hover-primary);
+		}
+	}
+}
+
 .stage {
+	width: 100vw;
+	overflow: hidden;
+	transition: width $animation-duration-width $animation-easing-width;
+	transition-delay: $animation-duration - $animation-duration-width;
+	text-align: center;
+
 	background-color: var(--skin-color-primary, $primary);
 	color: var(--text-color-contrast-switch, $text-invert);
 	flex-direction: column;
@@ -15,25 +49,71 @@ $gridSpacing: $grid-base-four;
 		display: flex;
 	}
 
+	&--no-animation {
+		transition: none;
+
+		.stage__title {
+			transition: none;
+		}
+	}
+
+	&--open {
+		left: -100%;
+
+		@include breakpoint($fromXLarge) {
+			left: 0;
+			width: 40vw;
+		}
+
+		.stage__title {
+			transform: none;
+		}
+	}
+
+	&--ready {
+		transition-delay: unset;
+	}
+
+	&__content,
+	&__loader {
+		flex-shrink: 0;
+		flex-grow: 0;
+	}
+
+	&__content {
+		display: flex;
+		flex-direction: column;
+		height: 100%;
+		align-self: center;
+	}
+
 	&__headline {
 		display: flex;
 		flex: 1 1 0;
 		flex-direction: column;
 		flex-wrap: wrap;
 		justify-content: flex-end;
-		margin-bottom: 140px;
 		word-break: normal;
 
-		h1 {
-			margin-bottom: $grid-base;
+		@include breakpoint($fromLarge) {
+			padding: $grid-base-six $grid-base-six 120px;
 		}
+	}
 
-		h4 {
-			margin: 0;
-			text-align: center;
-			font-size: $font-size-primary;
-			font-weight: $font-weight-regular;
+	&__title {
+		margin-bottom: $grid-base-four;
+
+		@include breakpoint($fromLarge) {
+			transform: scale(1.75);
+			transform-origin: bottom left;
+			transition: transform $animation-duration-title
+				$animation-easing-title;
+			transition-delay: #{$animation-duration - $animation-duration-title};
 		}
+	}
+
+	&__claim {
+		font-weight: $font-weight-regular;
 	}
 
 	&__logos {
@@ -46,7 +126,7 @@ $gridSpacing: $grid-base-four;
 		margin-bottom: $grid-base-five;
 		transform: scale(0.9);
 
-		@media (width >= calc(#{$xsmall} + #{$grid-base-three})) {
+		@media (width >=calc(#{$xsmall} + #{$grid-base-three})) {
 			transform: scale(1);
 			margin-bottom: $grid-base-ten;
 		}
@@ -84,56 +164,10 @@ $gridSpacing: $grid-base-four;
 		}
 	}
 
-	&--animated {
-		display: flex;
-		width: 100vw;
-
-		&.stage--animation-done {
-			display: none;
-
-			@include breakpoint($fromXLarge) {
-				display: flex;
-			}
-		}
-
-		&.stage--ready {
-			animation: overlayAnimation 1.4s ease forwards;
-			animation-delay: $loginOverlayDelayMobile;
-		}
-
-		.stage__headline {
-			margin-bottom: 80px;
-		}
-
-		@keyframes overlayAnimation {
-			from {
-				left: 0;
-			}
-			to {
-				left: -100vw;
-			}
-		}
-
-		@include breakpoint($fromXLarge) {
-			&.stage--ready {
-				animation: overlayAnimation 1s ease forwards;
-				animation-delay: $loginOverlayDelayDesktop;
-			}
-
-			@keyframes overlayAnimation {
-				from {
-					width: 100vw;
-				}
-				to {
-					width: 40vw;
-				}
-			}
-		}
-	}
-
 	&__spinner {
 		animation-delay: 2s;
 		animation: hideSpinner 4s ease forwards;
+		margin: 80px auto 0;
 
 		@include breakpoint($fromXLarge) {
 			animation-duration: 2s;
@@ -143,9 +177,57 @@ $gridSpacing: $grid-base-four;
 			from {
 				opacity: 1;
 			}
+
 			to {
 				opacity: 0;
 			}
 		}
+	}
+
+	&__legalLinks {
+		display: flex;
+		align-items: center;
+		align-self: center;
+		margin: 40px 0;
+
+		.button-as-link {
+			background: none;
+			border: none;
+			cursor: pointer;
+			padding: 0;
+			white-space: nowrap;
+			color: $tertiary;
+
+			&:focus {
+				outline: $focus-outline;
+				border-radius: 4px;
+			}
+
+			&:focus:not(:focus-visible) {
+				outline: none;
+			}
+
+			&:hover {
+				color: var(--skin-color-primary-hover, $hover-primary);
+			}
+		}
+
+		.text {
+			font-size: $font-size-tertiary;
+		}
+	}
+
+	&__legalLinksItem {
+		color: $dark-grey;
+
+		&:hover {
+			color: $white;
+		}
+	}
+
+	&__legalLinksSeparator {
+		display: inline-block;
+		color: $white;
+		margin: 0 $grid-base;
 	}
 }
