+++ src/components/stageLayout/StageLayout.tsx
@@ -22,7 +22,7 @@ import {
 	useScrollTrigger
 } from '@mui/material';
 import LoginIcon from '@mui/icons-material/Login';
-import { InfoDrawer } from '../../extensions/components/registration/infoDrawer/InfoDrawer';
+import { InfoDrawer } from '../registration/infoDrawer/InfoDrawer';
 
 interface StageLayoutProps {
 	className?: string;
