--- src/extensions/resources/styles/generateQrCode.scss
@@ -1,3 +1,0 @@
-.generateQrCode__overlayContent .text {
-	margin-left: 0;
-}
