--- src/extensions/resources/styles/releaseNote.styles.scss
@@ -1,3 +1,0 @@
-.releaseNote.overlay {
-	z-index: 9999;
-}
