--- src/extensions/resources/styles/twoFactorAuth.scss
@@ -1,4 +1,0 @@
-.twoFactorAuth__emailConfirmation,
-.twoFactorAuth__appConfirmation {
-	text-align: center;
-}
