+++ src/resources/styles/mui-variables-mapping.scss
@@ -14,7 +14,7 @@
 	--button-border-radius: #{$button-border-radius};
 	--hover-primary: #{$hover-primary};
 	--secondary: #{$secondary};
-	--primary-3: #{$primary-3};
+	--primary-3: #{$hover-primary};
 	--tertiary: #{$tertiary};
 	--black: #{$black};
 	--form-error: #{$form-error};
